Do not require 'enactment_date' of documents
Notices do not have an enactment_date, but are returned as documents.
This requires us to handle the case of documents without enactment_dates
gracefully. Still, the expectation that documents except for notices
have enactment_dates stands, but is not codified as such.

The addition of parentheses is also handled in the case of an
enactment_date being present, as otherwise there would be a set of empty
parentheses in the output.