Use filename as title fallback in File HTML
The title property has been renamed to filename, and a new title
property has been introduced with 1.2.0. This new property may be an
empty string, rendering the output useless.

By changing to a fallback order 'title -> filename' we can be sure to
always have a string in the <a> tag.