Revert to old usage of file attribute 'title'
In version 1.2.0 of the API, the file attribute 'title' contains the
filename. The additional attribute 'description' may contain a
user-defined title string associated with the file.

The 'filename' attribute has been removed from the API, which is
reflected in the changes to the XML examples and the XSD.

Graphically, the changes within this branch reflect the development
state of the API as follows:

  1.1.1       (WIP)        1.2.0
  ------------------------------------
  title   ->  filename  -> title
    -     ->  title     -> description

Regardless of the attribute naming scheme chosen in the XML returned
from the API, we handle the data semantically correct within the File
entity.